<==== Mercurial data ====>
Branch: dev_2.5

+++ mediaserver_core/src/recorder/storage_manager.cpp
@@ -1769,7 +1769,7 @@ QnStorageResourcePtr QnStorageManager::getOptimalStorageRoot(QnAbstractMediaStre
 
     struct StorageSpaceInfo {
         QnStorageResourcePtr storage;
-        qint64 usageCoeff;
+        double usageCoeff;
     };
     std::vector<StorageSpaceInfo> storagesInfo;
 
@@ -1786,7 +1786,7 @@ QnStorageResourcePtr QnStorageManager::getOptimalStorageRoot(QnAbstractMediaStre
                         .arg((*it)->getUrl())
                         .arg((*it)->calcUsageCoeff());
         if ((*it)->calcUsageCoeff() >= 0) {
-            StorageSpaceInfo tmp = {*it, static_cast<qint64>((*it)->calcUsageCoeff())};
+            StorageSpaceInfo tmp = {*it, (*it)->calcUsageCoeff()};
             storagesInfo.push_back(tmp);
         }
     }
